# turnos/contracts/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ChangePasswordRequest" : {
        "properties" : {
          "currentPassword" : {
            "minLength" : 1,
            "type" : "string"
          },
          "newPassword" : {
            "maxLength" : 200,
            "minLength" : 10,
            "type" : "string"
          }
        },
        "required" : [ "currentPassword", "newPassword" ],
        "type" : "object"
      },
...
      "LoginRequest" : {
        "properties" : {
          "email" : {
            "minLength" : 1,
            "type" : "string"
          },
          "password" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "email", "password" ],
        "type" : "object"
      },
      "MeResponse" : {
        "properties" : {
          "admin" : {
            "type" : "boolean"
          },
          "email" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "rank" : {
            "type" : "string"
          },
          "serviceNumber" : {
            "type" : "string"
          }
        },
        "required" : [ "admin", "email", "fullName", "id", "rank", "serviceNumber" ],
        "type" : "object"
      },
...
      "ProfileRequest" : {
        "properties" : {
          "fullName" : {
            "maxLength" : 120,
            "minLength" : 1,
            "type" : "string"
          },
          "rank" : {
            "maxLength" : 40,
            "minLength" : 0,
            "type" : "string"
          },
          "serviceNumber" : {
            "pattern" : "^([0-9]{1,5})?$",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RegisterRequest" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          },
          "fullName" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "inviteCode" : {
            "maxLength" : 40,
            "minLength" : 0,
            "type" : "string"
          },
          "password" : {
            "maxLength" : 200,
            "minLength" : 10,
            "type" : "string"
          },
          "rank" : {
            "maxLength" : 40,
            "minLength" : 0,
            "type" : "string"
          },
          "serviceNumber" : {
            "pattern" : "^[0-9]{1,5}$",
            "type" : "string"
          }
        },
        "required" : [ "email", "fullName", "password" ],
        "type" : "object"
      },
      "ResetRequest" : {
        "properties" : {
          "code" : {
            "maxLength" : 40,
            "minLength" : 0,
            "type" : "string"
          },
          "password" : {
            "maxLength" : 200,
            "minLength" : 10,
            "type" : "string"
          }
        },
        "required" : [ "code", "password" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/v1/auth/login" : {
      "post" : {
        "operationId" : "login",
        "parameters" : [ {
          "in" : "header",
          "name" : "User-Agent",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
...
    "/api/v1/auth/password-reset" : {
      "post" : {
        "operationId" : "resetPassword",
        "parameters" : [ {
          "in" : "header",
          "name" : "User-Agent",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResetRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
...
    "/api/v1/auth/register" : {
      "post" : {
        "operationId" : "register",
        "parameters" : [ {
          "in" : "header",
          "name" : "User-Agent",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
...
    "/api/v1/me" : {
      "get" : {
        "operationId" : "me",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth-controller" ]
      },
      "patch" : {
        "operationId" : "updateProfile",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProfileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
...
    "/api/v1/me/password" : {
      "post" : {
        "operationId" : "changePassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "url" : "/"
  } ]
}